requirements.txt

azure-core==1.32.0
azure-identity==1.20.0
docx2pdf
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.23
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langchain-unstructured==0.1.5
langsmith==0.1.147
openai==1.57.2
openpyxl==3.1.5
PyJWT==2.10.1
PyMuPDF==1.25.3
python-docx
pytz==2024.2
tiktoken==0.8.0
tinycss2==1.4.0
streamlit
pydantic==2.8
Spire.Doc
plum-dispatch==1.7.4
faiss-cpu
mammoth==1.5.1
reportlab==3.6.12